Use OHLC when ticks is not sufficient, refactor
