+++ .storybook/open-contracts.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import { storiesOf } from '@kadira/storybook';
-import DynamicChart from './DynamicChart';
+import ContractChart from './ContractChart';
+
+const contract5day = '8686428788';
 
 storiesOf('Open Contracts', module)
-    .add('Empty', () =>
-        <DynamicChart />
+    .add('5 days contract', () =>
+        <ContractChart contractID={contract5day} />
     );
